+++ src/SuffixAutomaton.h
@@ -6,19 +6,19 @@
 #include <string>
 
 struct State{
-	int len, link;
+	long long len, link;
 	long long first_pos = -1;
-	std::map<char, int> next;
+	std::map<char, long long> next;
 };
 
 extern std::vector<State> st;
-extern int sz;
-extern int last;
-extern std::vector<std::vector<int>> inv_link;
+extern long long sz;
+extern long long last;
+extern std::vector<std::vector<long long>> inv_link;
 
 void sa_init();
 void sa_extend(char c, long long current_pos);
 void build_tree();
-void get_occurences(int v, std::vector<long long>& positions);
+void get_occurences(long long v, std::vector<long long>& positions);
 
 #endif
